pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dlmi-hw1"
version = "0.1.0"
description = "DLMI HW1: BraTS medical image synthesis baseline"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "kagglehub>=0.3.3",
  "nibabel>=5.2.1",
  "monai>=1.3.0",
  "numpy>=1.26.0",
  "torch>=2.2.0",
  "torchvision>=0.17.0",
  "tqdm>=4.66.0",
  "scikit-image>=0.22.0",
  "pyyaml>=6.0.1",
  "matplotlib>=3.8.0",
]

[tool.setuptools.packages.find]
where = ["src"]
